[PIR #920] Extract webview markup/script to backlog-search.template.ts
